HT_22/requirements.txt
django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
requests==2.31.0